+++ dimers/ssfsevolve.cpp
@@ -3,7 +3,7 @@
 #include <omp.h>
 using namespace std;
 void ssfsevolve(double J1, int* states, int statesize, int* spinstates, int spinstatesize, int* s2p,
-  int nd,  double* walker2params, double* walker2ids, double* energies, int* failedupdates,
+  int nd,  double* walker2params, int* walker2ids, double* energies, int* failedupdates,
   int nbwalkers, int nthreads, int iters, int nt, int nh) {
     #pragma omp parallel for schedule(dynamic, 1) num_threads(nthreads)
     for(int walker = 0; walker < nbwalkers; walker++) {
